requirements.txt
torch>=2.0.0
torchvision>=0.15.0
safetensors>=0.4.0
diffusers>=0.29.0
transformers>=4.40.0
peft>=0.10.0
pyyaml>=6.0
tqdm>=4.65.0
rasterio>=1.3.9
numpy>=1.24.0
Pillow>=9.5.0
matplotlib>=3.7.0
scipy>=1.10.0
pandas>=2.0.0